+++ surbear-core/src/main/java/com/surbear/member/entity/MemberEntity.java
@@ -1,8 +1,8 @@
 package com.surbear.member.entity;
 
 import com.surbear.common.entity.BaseTimeEntity;
-import com.surbear.member.model.Age;
-import com.surbear.member.model.Gender;
+import com.surbear.member.constant.Age;
+import com.surbear.member.constant.Gender;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
